+++ docs/changes.md
@@ -13,6 +13,8 @@
 - Added camera component FieldOfView.
 - Fixed collision between static / kinematic objects and rigid bodies.
 - Fixed 'Destroy Object When Finish' + TTL on animators (deletion didn't work if TTL was longer than animation).
+- Added position offset to sprites and billboards.
+- Added calculating bounding box and bounding sphere to sprites and billboards.
 
 ## 0.1.0.7